iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/rest/request/alternative/CryptoBookRequestBuilderTest.java
package pl.zankowski.iextrading4j.client.rest.request.alternative;

import org.junit.jupiter.api.Test;
import pl.zankowski.iextrading4j.api.marketdata.Book;
import pl.zankowski.iextrading4j.client.rest.manager.MethodType;
import pl.zankowski.iextrading4j.client.rest.manager.RestRequest;

import javax.ws.rs.core.GenericType;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.entry;

class CryptoBookRequestBuilderTest {

    @Test
    void shouldSuccessfullyCreateRequest() {
        final String symbol = "BTCUSD";

        final RestRequest<Book> request = new CryptoBookRequestBuilder()
                .withSymbol(symbol)
                .build();

        assertThat(request.getMethodType()).isEqualTo(MethodType.GET);
        assertThat(request.getPath()).isEqualTo("/crypto/{symbol}/book");
        assertThat(request.getResponseType()).isEqualTo(new GenericType<Book>() {});
        assertThat(request.getPathParams()).containsExactly(entry("symbol", symbol));
        assertThat(request.getQueryParams()).isEmpty();
    }

}
